requirements.txt
# FastAPI & Web
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
python-multipart>=0.0.6
pydantic>=1.10.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0

# HTTP & Auth
httpx>=0.24.1
aiohttp>=3.9.5
PyJWT[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4

# Cache & Utils
redis>=4.5.0
python-dotenv>=1.0.0

# AI/ML Libraries
# openai>=1.0.0  # Not used, using Groq instead
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.2

# ML & Data Science
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# CS2 Demo Parsing (optional - requires Python 3.10+)
# demoparser2>=0.1.0
# awpy>=1.3.0

# Monitoring
prometheus-client>=0.19.0
